+++ src/main/java/net/minecraft/client/renderer/entity/Render.java
@@ -360,7 +360,7 @@ public abstract class Render<T extends Entity> implements IEntityRenderer, Entit
                 i = -10;
             }
 
-            int j = (int) (fontrenderer.getStringWidth(str) / 2);
+            int j = fontrenderer.getStringWidth(str) / 2;
             GlStateManager.disableTexture2D();
             worldrenderer.begin(7, DefaultVertexFormats.POSITION_COLOR);
             worldrenderer.pos(-j - 1, -1 + i, 0.0D).color(0.0F, 0.0F, 0.0F, 0.25F).endVertex();
