+++ src/main/java/net/minecraft/command/CommandTime.java
@@ -104,7 +104,7 @@ public class CommandTime extends CommandBase
     {
         for (int i = 0; i < MinecraftServer.getServer().worldServers.length; ++i)
         {
-            MinecraftServer.getServer().worldServers[i].setWorldTime((long)time);
+            MinecraftServer.getServer().worldServers[i].setWorldTime(time);
         }
     }
