+++ src/main/java/net/minecraft/entity/ai/EntityAIFleeSun.java
@@ -86,7 +86,7 @@ public class EntityAIFleeSun extends EntityAIBase
 
             if (!this.theWorld.canSeeSky(blockpos1) && this.theCreature.getBlockPathWeight(blockpos1) < 0.0F)
             {
-                return new Vec3((double)blockpos1.getX(), (double)blockpos1.getY(), (double)blockpos1.getZ());
+                return new Vec3(blockpos1.getX(), blockpos1.getY(), blockpos1.getZ());
             }
         }
