+++ src/main/java/net/minecraft/nbt/NBTTagByteArray.java
@@ -32,7 +32,7 @@ public class NBTTagByteArray extends NBTBase
     {
         sizeTracker.read(192L);
         int i = input.readInt();
-        sizeTracker.read((long)(8 * i));
+        sizeTracker.read(8 * i);
         this.data = new byte[i];
         input.readFully(this.data);
     }
